+++ xayn-ai/src/coi/stats.rs
@@ -1,0 +1,75 @@
+use std::time::{Duration, SystemTime};
+
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    coi::point::{NegativeCoi, PositiveCoi},
+    utils::system_time_now,
+};
+
+#[derive(Clone, Copy, Debug, Deserialize, Serialize)]
+pub(crate) struct CoiStats {
+    pub(crate) view_count: usize,
+    pub(crate) view_time: Duration,
+    pub(crate) last_view: SystemTime,
+}
+
+impl CoiStats {
+    pub(crate) fn new(viewed: Option<Duration>) -> Self {
+        Self {
+            view_count: 1,
+            view_time: viewed.unwrap_or_default(),
+            last_view: system_time_now(),
+        }
+    }
+
+    pub(crate) fn update(&mut self, viewed: Option<Duration>) {
+        self.view_count += 1;
+        if let Some(viewed) = viewed {
+            self.view_time += viewed;
+        }
+        self.last_view = system_time_now();
+    }
+
+    pub(crate) fn merge(self, other: Self) -> Self {
+        Self {
+            view_count: self.view_count + other.view_count,
+            view_time: self.view_time + other.view_time,
+            last_view: self.last_view.max(other.last_view),
+        }
+    }
+}
+
+impl Default for CoiStats {
+    fn default() -> Self {
+        Self {
+            view_count: 1,
+            view_time: Duration::ZERO,
+            last_view: SystemTime::UNIX_EPOCH,
+        }
+    }
+}
+
+pub(crate) trait CoiPointStats {
+    fn stats(&self) -> CoiStats;
+
+    fn update_stats(&mut self, viewed: Option<Duration>);
+}
+
+impl CoiPointStats for PositiveCoi {
+    fn stats(&self) -> CoiStats {
+        self.stats
+    }
+
+    fn update_stats(&mut self, viewed: Option<Duration>) {
+        self.stats.update(viewed);
+    }
+}
+
+impl CoiPointStats for NegativeCoi {
+    fn stats(&self) -> CoiStats {
+        CoiStats::default()
+    }
+
+    fn update_stats(&mut self, _viewed: Option<Duration>) {}
+}
